Refactored MovingHandler into their own classes.
